better split of services, merging both api implementations
