+++ include/web_task.h
@@ -9,9 +9,6 @@ extern "C" {
 
 #include <Arduino.h>
 
-#define MAX_IR_TEXT_CODE_LENGTH 2048
-#define MAX_IR_FORMAT_TYPE 50
-
 void TaskWeb(void *pvParameters);
 
 #ifdef __cplusplus
