+++ platformio.ini
@@ -12,6 +12,13 @@
 default_envs = esp32doit
 
 [common]
+lib_deps = 
+	bt_service
+	api_service
+	config
+	mdns
+	ir_service
+
 lib_deps_external = 
 	me-no-dev/AsyncTCP @ ^1.1.1
 	https://github.com/me-no-dev/ESPAsyncWebServer.git#f71e3d427b5be9791a8a2c93cf8079792c3a9a26
@@ -24,6 +31,8 @@ board = az-delivery-devkit-v4
 framework = arduino
 lib_deps =
   ${common.lib_deps_external}
+  ${common.lib_deps}
+
 board_build.partitions = partitions.csv
 monitor_speed = 115200
 
@@ -33,6 +42,8 @@ board = esp32doit-devkit-v1
 framework = arduino
 lib_deps =
   ${common.lib_deps_external}
+  ${common.lib_deps}
+  
 board_build.partitions = partitions.csv
 monitor_speed = 115200
-build_flags = -DBLASTER_PIN_INDICATOR=2 
+build_flags = -DBLASTER_PIN_INDICATOR=2 -DBLASTER_PIN_IR_INTERNAL=13 -DBLASTER_PIN_IR_OUT_1=12
