+++ src/ir_task.cpp
@@ -6,8 +6,10 @@
 
 #include <ir_message.h>
 #include <ir_task.h>
-#include <blaster_config.h>
 #include <ir_queue.h>
+
+#include "blaster_config.h"
+
 
 uint16_t irRepeat = 0;
 ir_message_t repeatMessage;
@@ -54,7 +56,7 @@ void sendHexCode(ir_message_t &message)
         irsend.send(message.decodeType, message.code64, message.codeLen, irRepeat);
     }
 }
-
+//int blub=0;
 void TaskSendIR(void *pvParameters)
 {
     Serial.printf("TaskSendIR running on core %d\n", xPortGetCoreID());
@@ -74,6 +76,12 @@ void TaskSendIR(void *pvParameters)
                 {
                     uint32_t ir_pin_mask = 0 | 1 << BLASTER_PIN_INDICATOR | message.ir_internal << BLASTER_PIN_IR_INTERNAL | message.ir_ext1 << BLASTER_PIN_IR_OUT_1 | message.ir_ext2 << BLASTER_PIN_IR_OUT_2;
                     irsend.setPinMask(ir_pin_mask);
+//                    blub += 1;
+//                    if (blub%2 == 1){
+//                        GPIO.out_w1ts = ir_pin_mask;
+//                    } else {
+//                        GPIO.out_w1tc = ir_pin_mask;
+//                    }
                     switch (message.format)
                     {
                     case pronto:
